refactor(crates): move SplitReason into media-types
SplitReason, VideoCodecInfo and AudioCodecInfo now live in the dependency-free
media-types crate; pipeline-common re-exports them from split_reason.rs so its
downstream users (flv-fix, hls-fix, mesio, the app) keep importing via
pipeline_common unchanged. flv and hls import the types from media-types and no
longer depend on pipeline-common, which they used solely for SplitReason.
