+++ crates/flv/Cargo.toml
@@ -19,7 +19,7 @@ amf0 = { path = "../amf0" }
 av1 = { path = "../av1" }
 h264 = { path = "../h264" }
 h265 = { path = "../h265" }
-pipeline-common = { path = "../pipeline-common" }
+media-types = { path = "../media-types" }
 thiserror = { workspace = true }
 tracing = { workspace = true }
