+++ crates/flv/src/lib.rs
@@ -22,7 +22,7 @@ pub mod writer_async;
 pub use data::FlvData;
 pub use error::FlvError;
 pub use header::FlvHeader;
-pub use pipeline_common::split_reason::{AudioCodecInfo, SplitReason, VideoCodecInfo};
+pub use media_types::split_reason::{AudioCodecInfo, SplitReason, VideoCodecInfo};
 pub use tag::{CodecKind, FlvTag, FlvTagType, TagClass};
 pub use writer::FlvWriter;
 pub use writer_async::FlvEncoder;
