+++ crates/hls/Cargo.toml
@@ -17,6 +17,5 @@ mp4 = { path = "../mp4" }
 h264 = { path = "../h264" }
 h265 = { path = "../h265" }
 media-types = { path = "../media-types" }
-pipeline-common = { path = "../pipeline-common" }
 tracing = { workspace = true }
 memchr = { workspace = true }
