+++ crates/pipeline-common/Cargo.toml
@@ -10,6 +10,7 @@ license.workspace = true
 workspace = true
 
 [dependencies]
+media-types = { path = "../media-types" }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 time = { workspace = true, features = ["local-offset"] }
